implemented note creation and editing to frontend

+++ frontend/src/scenes/Notes/components/Note.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { gql } from 'apollo-boost'
 import { useMutation } from '@apollo/react-hooks'
 import { SCHOOL_NOTES } from '../'
@@ -14,21 +14,80 @@ const DELETE_NOTE = gql`
 	}
 `
 
+const EDIT_NOTE = gql`
+	mutation EditSchoolNote($id: Int!, $header: String!, $content: String!) {
+		editSchoolNote(
+			id: $id,
+			header: $header,
+			content: $content
+		){
+			success,
+			id
+		}
+	}
+`
+
 const Note = ({ note }) => {
+	const [editmode, setEditMode] = useState(false)
+	const [newHeader, setNewHeader] = useState(note.header)
+	const [newContent, setNewContent] = useState(note.content)
 	const [deleteNote] = useMutation(DELETE_NOTE, {
+		refetchQueries: [{ query: SCHOOL_NOTES }],
+	})
+	const [editNote] = useMutation(EDIT_NOTE, {
 		refetchQueries: [{ query: SCHOOL_NOTES }],
 	})
 
 	const removeNote = async () => {
-		await deleteNote({ variables: { id: note.id } })
+		const message = `Do you really want to delete '${note.header}'`
+		if (window.confirm(message)) {
+			await deleteNote({ variables: { id: note.id } })
+		}
+	}
+
+	const cancel = async () => {
+		setEditMode(false)
+		setNewHeader(note.header)
+		setNewContent(note.content)
+	}
+
+	const save = async () => {
+		await editNote({
+			variables: {
+				id: note.id,
+				header: newHeader,
+				content: newContent,
+			},
+		})
 	}
 
 	return (
 		<div style={{ backgroundColor: 'yellow' }}>
-			<h2>{note.header}</h2>
-			Subjects: {note.subjects.map(s => s.name)}<br />
-			Courses: {note.subjects.map(s => s.courses.map(c => c.name))}
-			<p>{note.content}</p>
+			{!editmode
+				? <h2>{note.header}</h2>
+				: <><br /><input
+					type='text'
+					value={newHeader}
+					onChange={({ target }) => setNewHeader(target.value)}
+				></input><br /></>
+			}
+			Subjects: {note.subjects.map(s => `${s.name} `)}<br />
+			Courses: {note.subjects.map(s => s.courses.map(c => `${c.name} `))}
+			{!editmode
+				? <p>{note.content}</p>
+				: <><br /><textarea
+					rows='6' cols='50'
+					value={newContent}
+					onChange={({ target }) => setNewContent(target.value)}
+				></textarea><br /></>
+			}
+			{!editmode
+				? <button onClick={() => setEditMode(true)}>Edit</button>
+				: <>
+					<button onClick={save}>Save</button>
+					<button onClick={cancel}>Cancel</button>
+				</>
+			}
 			<button onClick={removeNote}>DELETE</button>
 		</div>
 	)
